+++ mer-admin-panel/src/redux/api/apiSlice.ts
@@ -116,6 +116,8 @@ export const apiSlice = createApi({
     "AllCoupons",
     "Coupon",
     "AllStaff",
-    "Stuff"
+    "Stuff",
+    "AllCollections",
+    "Collection"
   ],
 });